@@ -1,35 +1,36 @@
-// ChatForm.js
 import React, { useState } from 'react';
-import './ChatForm.css'; // Import the corresponding CSS file
+import './ChatForm.css'; // Ensure this CSS file is in the same directory
 
-const ChatForm = ({onSubmit}) => {
+const ChatForm = ({ onSubmit }) => {
   const [name, setName] = useState('');
 
-  const handleNameChange = (e) => {
+  // Function to handle input change
+  const handleInputChange = (e) => {
     setName(e.target.value);
   };
 
+  // Function to handle form submission
   const handleSubmit = (e) => {
-    e.preventDefault();
-    onSubmit(e, name);
-    setName('');
-    console.log('Submitted name:', name);
+    e.preventDefault(); // Prevent default form submission behavior
+    if (name.trim()) { // Check if the input is not empty
+      onSubmit(name); // Pass the input value to the parent component
+      setName(''); // Clear the input field
+    }
   };
 
   return (
-    <div className="chat-form-container">
-      <form onSubmit={handleSubmit}>
-        <label htmlFor="nameInput">What's your name?</label>
-        <input
-          type="text"
-          id="nameInput"
-          placeholder="Type your name here..."
-          value={name}
-          onChange={handleNameChange}
-        />
-        <button type="submit">Send</button>
-      </form>
-    </div>
+    <form onSubmit={handleSubmit} className="chat-form-container">
+      <label htmlFor="nameInput">What's your name?</label>
+      <input
+        type="text"
+        id="nameInput"
+        placeholder="Type your name here..."
+        value={name}
+        onChange={handleInputChange}
+        required
+      />
+      <button type="submit">Send</button>
+    </form>
   );
 };
 
